Refactor code copying mechanism
